tests/wpt/web-platform-tests/css/css-ruby/parsing/ruby-align-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Ruby Layout: parsing ruby-align with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-ruby-1/#ruby-align-property">
<meta name="assert" content="ruby-align supports the full grammar 'start | center | space-between | space-around'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("ruby-align", "start");
test_valid_value("ruby-align", "center");
test_valid_value("ruby-align", "space-between");
test_valid_value("ruby-align", "space-around");
</script>
</body>
</html>
